add: prices

+++ database/migrations/2024_05_17_152150_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->unsignedBigInteger('ean13')->unique();
             $table->string('name')->unique();
             $table->string('slug')->unique();
-            $table->float('price');
-            $table->float('price_with_discount')->nullable();
+            $table->integer('price');
+            $table->integer('price_with_discount')->nullable();
             $table->boolean('published')->default(false)->index();
             $table->unsignedInteger('stock');
             $table->float('dimension_length');
